fix(grammar): fix grammar conflicts and refactor
The grammar had some conflicts, which were resolved by adding precedence
to qualident and designator.

The scanner was throwing some errors related to spaces, and it was
required to add the spaces to the extras option.

Refactored the grammar, by adding module_header and module_footer rules.

Finally, string regular expression was missing a * that would accept
more than one character. The string and integer regular tokens were also
added to the grammar rules.