fix: add qualident option to type_decl
type_decl can be either a qualident, or a struct_type. Somehow, the
grammar I had implemented before, did not have this. As a result, I had
to fix it and add the missing qualident. Otherwise, it would only
recognise struct types such as arrays, procedure types, records, etc...

After fixing the rule, the parser files were regenerated, and
subsequently added to the repository.