refactor: replace occurrences of '-', '+', '/', '*', and 'END' by keyword definitions
In some places of the grammar, the mathematical operators are needed in
order to differentiate export privacy and expressions. Therefore, all
ocurrences of the mathematical operators were replaced by token
definitions. This means that in the AST we need to declare them as well.
The same goes for the END keyword, that in the if statement had the
literal 'END' as opposed to the keyword.