chore: update package for Oberon-07
When I implemented this grammar, it was meant to be a parser for
Oberon-2. However, the grammar I followed was that of Oberon-7. So,
instead of modifying the grammar for Oberon-2, I'll work on it, improve
it and then make a fork that will be used as a basis for Oberon-2, and
subsequently Component Pascal.