feat: make semicolon in last field declaration of a record optional
The standard for Oberon-7 does not allow the last field declaration of a
record type to be terminated by a semicolon. This was implemented
correctly in the parser. However, the majority of the compilers, and
that includes Wirth's own compiler make this optional. Hence, to be able
to handle those compilers, I have also made it optional.